shanawar test case

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -18,7 +18,8 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            //Because the application is have financial term related when even i single cent is important so that why float is nor perferable as it can effect precission value,So we use decimal
+            $table->decimal('commission_rate', 10, 4);
             $table->string('discount_code');
             $table->timestamps();
         });
